email notif on account request, implementd admin account creation

+++ server/api/auth/user.post.ts
@@ -4,21 +4,24 @@ const prisma = new PrismaClient();
 
 export default defineEventHandler(async (event) =>{
     const body = await readBody(event);
-    console.log(body)
     await prisma.user.create({
         data:{
              name: body.name,
              email: body.email,
+             permission: body.permission,
              id: body.id,
          }
         
      })
     console.log("Account Created");
-    await prisma.request.delete({
+    if(body.isRequest){
+        await prisma.request.delete({
         where:{
             id:body.id,
         }
-    })
-    console.log("request deleted");
+        });
+        console.log("request deleted");
+    }
+    
 });
